// angular/src/services/interfaces/models.ts
export interface Contact {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    customerId?: number;
    customer?: Customer;
    locationId?: number | undefined;
    location?: Location;
    positionTypeId?: number | undefined;
    positionType?: PositionType;
    contactName?: string | undefined;
    phoneNumber?: string | undefined;
    email?: string | undefined;
    notes?: string | undefined;
    isEmailOrder?: boolean;
    isEmailCreditMemo?: boolean;
    isEmailStatement?: boolean;
}

export interface CreateOrderResult {
    order?: Order;
    orderNumber?: number | undefined;
    quoteNumber?: number | undefined;
    orderResult?: boolean;
    syncResult?: boolean | undefined;
    isQuote?: boolean;
}
// ...
export interface CustomerCredit {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    customerId?: number;
    customer?: Customer;
    referenceId?: number | undefined;
    creditType?: number;
    creditReason?: string | undefined;
    previousRecordId?: number;
    amountPostedDate?: moment.Moment;
    currentBalance?: number;
    previousBalance?: number;
    amountPosted?: number;
}

export interface CustomerDTO {
    id?: number;
    accountNumber?: number;
    customerName?: string | undefined;
    discount?: number | undefined;
    taxRate?: number | undefined;
    paymentTermId?: number | undefined;
    priceLevelId?: number | undefined;
    addressLine1?: string | undefined;
    addressLine2?: string | undefined;
    city?: string | undefined;
    state?: string | undefined;
    country?: string | undefined;
    zipCode?: string | undefined;
    zone?: string | undefined;
    contactName?: string | undefined;
    phoneNumber?: string | undefined;
    faxNumber?: string | undefined;
    email?: string | undefined;
    isHoldAccount?: boolean;
    creditLimit?: number;
    isBypassCreditLimit?: boolean;
    overBalance?: number;
    statementAmount?: number;
    contacts?: Contact[] | undefined;
    customerCredit?: CustomerCredit;
    locations?: Location[] | undefined;
}
// ...
export interface Location {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    addressLine1?: string | undefined;
    addressLine2?: string | undefined;
    city?: string | undefined;
    state?: string | undefined;
    country?: string | undefined;
    zipCode?: string | undefined;
    phoneNumber?: string | undefined;
    faxNumber?: string | undefined;
    email?: string | undefined;
    latitude?: string | undefined;
    longitude?: string | undefined;
    customerId?: number;
    customer?: Customer;
    locationTypeId?: number;
    locationType?: LocationType;
    locationName?: string | undefined;
    locationCode?: string | undefined;
    notes?: string | undefined;
    zone?: string | undefined;
}
// ...
export interface LocationType {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    name?: string | undefined;
    code?: string | undefined;
    description?: string | undefined;
    sortOrder?: number;
}
// ...
export interface Order {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    customerId?: number;
    customer?: Customer;
    paymentTermId?: number | undefined;
    paymentTerm?: PaymentTerm;
    warehouseId?: number;
    warehouse?: Warehouse;
    orderNumber?: number | undefined;
    quoteNumber?: number | undefined;
    customerName?: string | undefined;
    accountNumber?: number;
    paymentTermName?: string | undefined;
    priceLevelId?: number;
    priceLevelName?: string | undefined;
    warehouseName?: string | undefined;
    orderDate?: moment.Moment;
    user?: string | undefined;
    orderStatusId?: number;
    orderStatusName?: string | undefined;
    orderedBy?: string | undefined;
    orderedByEmail?: string | undefined;
    orderedByPhoneNumber?: string | undefined;
    orderedByNotes?: string | undefined;
    deliveryType?: number;
    deliveryDate?: moment.Moment;
    deliveryRoute?: number | undefined;
    discount?: number | undefined;
    restockingFee?: number | undefined;
    restockingAmount?: number | undefined;
    taxRate?: number | undefined;
    totalTax?: number | undefined;
    subTotalAmount?: number | undefined;
    totalAmount?: number | undefined;
    isQuote?: boolean;
    isPrinted?: boolean;
    isCustomerOrder?: boolean;
    linkedInvoiceNumber?: string | undefined;
    orderDetails?: OrderDetail[] | undefined;
    email?: string | undefined;
    phoneNumber?: string | undefined;
    billAddress?: string | undefined;
    billCity?: string | undefined;
    billState?: string | undefined;
    billZipCode?: string | undefined;
    billZone?: string | undefined;
    billPhoneNumber?: string | undefined;
    billContactName?: string | undefined;
    shipAddressName?: string | undefined;
    shipAddress?: string | undefined;
    shipCity?: string | undefined;
    shipState?: string | undefined;
    shipZipCode?: string | undefined;
    shipZone?: string | undefined;
    shipPhoneNumber?: string | undefined;
    shipContactName?: string | undefined;
    isHoldInvoice?: boolean;
    invoiceNumber?: number | undefined;
    fwd?: number | undefined;
    originalInvoiceNumber?: number | undefined;
    orderType?: number | undefined;
    handlingType?: number | undefined;
    creditType?: number | undefined;
    singleInvoice?: number | undefined;
    purchaseOrderNumber?: string | undefined;
    type?: string | undefined;
    category?: string | undefined;
    carrier?: string | undefined;
    trackingNumber?: string | undefined;
    salesRepresentative1?: string | undefined;
    salesRepresentative2?: string | undefined;
    invoiceNotes?: string | undefined;
    particular?: string | undefined;
    paymentReference?: string | undefined;
    postedBy?: string | undefined;
    accountingPaidBy?: string | undefined;
    transactionDate?: moment.Moment | undefined;
    postDate?: moment.Moment | undefined;
    paymentDate?: moment.Moment | undefined;
    debit?: number | undefined;
    credit?: number | undefined;
    quantity?: number | undefined;
    quantityShipped?: number | undefined;
    currentCost?: number | undefined;
    amountPaid?: number | undefined;
    balance?: number | undefined;
    payment?: number | undefined;
    listPrice?: number | undefined;
    rgaType?: number;
    rgaReason?: number;
    rgaReasonNotes?: string | undefined;
}

export interface OrderDetail {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    orderId?: number;
    order?: Order;
    productId?: number;
    product?: Product;
    categoryId?: number;
    orderQuantity?: number;
    onHandQuantity?: number | undefined;
    location?: string | undefined;
    warehouseLocationId?: number | undefined;
    partNumber?: string | undefined;
    partDescription?: string | undefined;
    brand?: string | undefined;
    mainPartsLinkNumber?: string | undefined;
    mainOEMNumber?: string | undefined;
    vendorCode?: string | undefined;
    vendorPartNumber?: string | undefined;
    vendorPrice?: number;
    vendorOnHand?: number;
    yearFrom?: number;
    yearTo?: number;
    listPrice?: number;
    wholesalePrice?: number;
    discountedPrice?: number;
    totalAmount?: number;
    restockingFee?: number | undefined;
    restockingAmount?: number | undefined;
    discount?: number;
    discountAmount?: number;
    discountRecord?: number;
    discountRecordAmount?: number;
    partsLinks?: string | undefined;
    oeMs?: string | undefined;
    vendorCodes?: string | undefined;
    imageUrl?: string | undefined;
    stocks?: WarehouseStock[] | undefined;
    vendorCatalogs?: VendorCatalog[] | undefined;
    warehouseTrackings?: WarehouseTracking[] | undefined;
    vehicle?: string | undefined;
    salesRepresentative?: string | undefined;
    warehouseLocation?: string | undefined;
    vendorInfo?: string | undefined;
    warehouseTracking?: string | undefined;
    carrier?: string | undefined;
    trackingNumber?: string | undefined;
    salesOrderNumber?: number | undefined;
    originalInvoiceNumber?: number | undefined;
    buyOutOrder?: number | undefined;
    partSize?: number | undefined;
    statusId?: number | undefined;
    shippedQuantity?: number | undefined;
    scannedQuantity?: number | undefined;
    pickedQuantity?: number | undefined;
    returnQuantity?: number | undefined;
    remainingQuantity?: number | undefined;
    itemCatalog?: number | undefined;
    toShip?: number | undefined;
    price?: number | undefined;
    unitCost?: number | undefined;
    shipCharge?: number | undefined;
    shipDate?: moment.Moment | undefined;
    rgaInspectedCode?: number | undefined;
    rgaState?: number | undefined;
    rgaLocation?: string | undefined;
    rgaPartNumber?: string | undefined;
    isCreditMemoCreated?: boolean;
}
// ...
export interface OrderPaginatedListDTO {
    recordCount?: number;
    data?: Order[] | undefined;
}
// ...
export interface TotalSalesDTO {
    salesAmount?: number;
    creditAmount?: number;
    netSalesAmount?: number;
    unitCost?: number;
    profitMargin?: number;
    returnPercentage?: number;
    salesPerson?: string | undefined;
}

export interface TotalVendorSalesDTO {
    vendorName?: string | undefined;
    vendorCode?: string | undefined;
    salesAmount?: number;
    quantity?: number;
}

export interface User {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    roleId?: number;
    role?: Role;
    customerId?: number | undefined;
    customer?: Customer;
    locationId?: number | undefined;
    location?: Location;
    userName?: string | undefined;
    email?: string | undefined;
    isCustomerUser?: boolean;
    isActivated?: boolean;
    isChangePasswordOnLogin?: boolean;
    passwordHash?: string | undefined;
    passwordSalt?: string | undefined;
}
// ...
export interface VendorCatalog {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    vendorCode?: string | undefined;
    vendorPartNumber?: string | undefined;
    partsLinkNumber?: string | undefined;
    price?: number;
    onHand?: number;
    cutoffTime?: moment.Moment;
}
// ...
export interface WarehouseStock {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    productId?: number;
    product?: Product;
    warehouseId?: number;
    warehouse?: Warehouse;
    warehouseLocationId?: number;
    warehouseLocation?: WarehouseLocation;
    quantity?: number;
    location?: string | undefined;
    currentCost?: number;
}